# PERF: faster dataframe construction from recarray