--- a/pandas/core/internals/construction.py
+++ b/pandas/core/internals/construction.py
@@ -176,7 +176,7 @@ def rec_array_to_mgr(
     # essentially process a record array then fill it
     fdata = ma.getdata(data)
     if index is None:
-        index = _get_names_from_index(fdata)
+        index = default_index(len(data))
     else:
         index = ensure_index(index)