- [x] closes #44826
- [ ] tests added / passed
- [x] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit) for how to run them
- [x] whatsnew entry

```python
n = int(1e6)
df = pd.DataFrame({"A": [.0]*n})
arr = df.to_records(index=False)

In [1]: %timeit pd.DataFrame(arr)
3.64 s ± 77.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each) <- master
2.23 ms ± 421 µs per loop (mean ± std. dev. of 7 runs, 100 loops each) <- PR
```